Refactored Sass and CSS tasks
